// apps/web/src/shared/api/gen/model/getOrder200Data.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Checkout API
 * POST /api/sessions с телом {} создаёт гостевую сессию. Вставьте data.token в Authorize без слова Bearer. Суммы указаны в копейках. Данные сохраняются после перезапуска. Оплата тестовая.
 * OpenAPI spec version: 1.0.0
 */
import type { GetOrder200DataCurrency } from './getOrder200DataCurrency';
import type { GetOrder200DataCustomer } from './getOrder200DataCustomer';
import type { GetOrder200DataDelivery } from './getOrder200DataDelivery';
import type { GetOrder200DataItemsItem } from './getOrder200DataItemsItem';
import type { GetOrder200DataPaymentMethod } from './getOrder200DataPaymentMethod';
import type { GetOrder200DataPaymentStatus } from './getOrder200DataPaymentStatus';
import type { GetOrder200DataStatus } from './getOrder200DataStatus';

export type GetOrder200Data = {
  id: string;
  /**
   * @minLength 1
   * @maxLength 500
   */
  number: string;
  status: (typeof GetOrder200DataStatus)[keyof typeof GetOrder200DataStatus];
  paymentStatus: (typeof GetOrder200DataPaymentStatus)[keyof typeof GetOrder200DataPaymentStatus];
  paymentMethod: (typeof GetOrder200DataPaymentMethod)[keyof typeof GetOrder200DataPaymentMethod];
  customer: GetOrder200DataCustomer;
  items: GetOrder200DataItemsItem[];
  delivery: GetOrder200DataDelivery;
  /**
   * Копейки; 129000 = 1 290 ₽.
   * @minimum 0
   */
  subtotal: number;
  /**
   * Копейки; 129000 = 1 290 ₽.
   * @minimum 0
   */
  shipping: number;
  /**
   * Копейки; 129000 = 1 290 ₽.
   * @minimum 0
   */
  total: number;
  currency: GetOrder200DataCurrency;
  createdAt: string;
};